flask/flask_modularizado/requirements.txt
blinker==1.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
emoji==0.6.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-MySQL==1.5.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonify==0.5
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.4
protobuf==3.14.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.0
rfc3986==1.4.0
schedule==0.6.0
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
sniffio==1.2.0
threadpoolctl==2.1.0
tqdm==4.52.0
urllib3==1.26.2
websocket-client==0.48.0
Werkzeug==1.0.1
Flask-Mail==0.9.1
uwsgi==2.0.19.1
flask-testing==0.8.1
